minor refactoring finished chat function calls added test for chatparser added dice token property type

+++ opengametable/src/net/rptools/maptool/client/ui/commandpanel/ChatExecutor.java
@@ -243,7 +243,7 @@ public class ChatExecutor {
 
 	private static String printRoll(DiceExpressionPart diceExpressionPart) {
 		DiceExpression diceExpression = diceExpressionPart.getDiceExpression();
-		int result=diceExpression.evaluate(new Random());
+		float result=diceExpression.evaluate(new Random());
 		StringBuilder sb=new StringBuilder();
 		sb.append(diceExpression.toString())
 			.append(" = ")
